backend/requirements.txt
flask
flask-cors
pymongo
pillow
pytesseract
pyjwt
